Move ColumnReader inside VectorizedParquetReader
